+++ NEWS.md
@@ -1,2 +1,4 @@
 # microhaplot 1.0.0
+* added new parameter for `prepHaplotFiles`: `n.jobs`. For any non-window OS, you can specific the number of SAM files to be parallel processed. We recommend two times the number of processors/cores. (9/18/19)
 * Introduces 3 main functions: `prepHaplotFiles`, `runShinyHaplot`, `mvShinyHaplot`
+